Fix bug when passing only one token to insert()
This patch fixes a bug which was discovered by a unchecked `unwrap()`
call which paniced when passing only one token to the `insert()`
function.

Now we have fixed this bug which was an unchecked edge-case: If the
passed token is the last one, we do not have to call the traversal
algorithm but can directly work with the passed value, safing a bit of
processing time. Therefore, a check was introduced to decide what values
and tokens to work with.

Unfortunately the tests didn't check this case before.

The version was bumped for this fix.